pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "AI_chat_memo"
version = "0.1.1"
description = "Учебный проект асинхронного чата с AI, извлекающего и использующего факты о пользователе для персонализированного контекста беседы"
authors = [{ name = "MaxSummerX" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "celery>=5.6.2",
    "email-validator>=2.3.0",
    "fastapi>=0.120.0",
    "gevent>=25.9.1",
    "greenlet>=3.2.4",
    "langchain-neo4j>=0.6.0",
    "loguru>=0.7.3",
    "mem0ai>=1.0.1",
    "ollama>=0.6.0",
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "redis>=7.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.20.0",
    "mypy>=1.18.2",
    "ruff>=0.15.0",
    "pre-commit>=4.3.0",
    "types-aiofiles",
    "celery-types>=0.24.0",
    "pytest-mock>=3.15.1",
    "bandit>=1.8.3"
]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = ["venv", ".venv", "alembic", "docker", ".git", "__pycache__", "app/migrations/*"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ASYNC",# asyncio
]
ignore = ["E203", "E266", "E501", "B008"]

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
exclude = [
    '\.venv',
    'venv',
    '__pycache__',
    'alembic/',
    'app/migration/'
]

[[tool.mypy.overrides]]
module = "mem0.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
